move content in inc/firmware/common to inc/common/firmware
